BACKPORT: xarray: Add XArray load operation
The xa_load function brings with it a lot of infrastructure; xa_empty(),
xa_is_err(), and large chunks of the XArray advanced API that are used
to implement xa_load.

As the test-suite demonstrates, it is possible to use the XArray functions
on a radix tree.  The radix tree functions depend on the GFP flags being
stored in the root of the tree, so it's not possible to use the radix
tree functions on an XArray.

Change-Id: I5d988e5d37279725414f70cf0a53f4a9827fbd6b
Signed-off-by: Matthew Wilcox <willy@infradead.org>