BACKPORT: page cache: Convert hole search to XArray
The page cache offers the ability to search for a miss in the previous or
next N locations.  Rather than teach the XArray about the page cache's
definition of a miss, use xas_prev() and xas_next() to search the page
array.  This should be more efficient as it does not have to start the
lookup from the top for each index.

Change-Id: I0c49a4b169b48e8e0a16256b4116c848039a48f4
Signed-off-by: Matthew Wilcox <willy@infradead.org>