BACKPORT: xarray: Step through an XArray
The xas_next and xas_prev functions move the xas index by one position,
and adjust the rest of the iterator state to match it.  This is more
efficient than calling xas_set() as it keeps the iterator at the leaves
of the tree instead of walking the iterator from the root each time.

Change-Id: I7df653749694e4ed19606da8c8ced1167827ede0
Signed-off-by: Matthew Wilcox <willy@infradead.org>