BACKPORT: xarray: Add xas_for_each_conflict
This iterator iterates over each entry that is stored in the index or
indices specified by the xa_state.  This is intended for use for a
conditional store of a multiindex entry, or to allow entries which are
about to be removed from the xarray to be disposed of properly.

Change-Id: Ia22de24f38b807db2fa2c7037991a85a531aa382
Signed-off-by: Matthew Wilcox <willy@infradead.org>